Centralise la construction des overlays et ajuste la scène au fond
